changed lmdmapdata ttree branch structure to resolve large file size problem, but without success. trying root file splitting now...

+++ data/PndLmdMapData.h
@@ -28,7 +28,7 @@ struct Point2D {
 
 struct Point2DCloud {
   std::map<Point2D, unsigned int> points;
-  unsigned int total_count;
+  unsigned long total_count;
 };
 
 class PndLmdMapData: public PndLmdAbstractData  {
